2.0 Release
git-svn-id: https://plugins.svn.wordpress.org/gallery-bank/trunk@742394 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ lib/album-gallery-bank-class.php
@@ -12,89 +12,35 @@ else
 	{
 		if($_REQUEST["param"] == "add_new_album")
 		{
-			$ux_album_name = html_entity_decode($_REQUEST["album_name"]);
+			$ux_album_name = html_entity_decode(esc_attr($_REQUEST["album_name"]));
 			$ux_desciption = html_entity_decode(esc_attr($_REQUEST["ux_description"]));
-			$ux_image_width = 160;
-			$ux_image_height = 120;
-			$ux_border_width = 5;
-			$ux_border_color = "#000";
-			$ux_image_no = 3;
-			$ux_slide_interval = 2;
-			$ux_slide = 0;
-			$ux_border = 1;
-			$images_in_row = 3;
+			$cover = esc_attr($_REQUEST["cover_array"]);
 			$wpdb->query
 			(
 				$wpdb->prepare
 				(
-					"INSERT INTO ".gallery_bank_albums()."(album_name,description,image_width,image_height,border_enable,
-					border_width,border_color,slideshow,slideshow_interval,album_date,author,images_in_row)
-					VALUES(%s,%s,%d,%d,%d,%d,%s,%d,%d,CURDATE(),%s,%d)",
+					"INSERT INTO ".gallery_bank_albums()."(album_name, description, album_date, author)
+					VALUES(%s, %s, CURDATE(), %s)",
 					$ux_album_name,
 					$ux_desciption,
-					$ux_image_width,
-					$ux_image_height,
-					$ux_border,
-					$ux_border_width,
-					$ux_border_color,
-					$ux_slide,
-					$ux_slide_interval,
-					$current_user->display_name,
-					$images_in_row
+					$current_user->display_name
 				)
 			);
 			echo $EventLastId=$wpdb->insert_id;
-			
-			die();
-		}
-		elseif($_REQUEST["param"] == "update_album")
-		{
-			$albumId = intval($_REQUEST["albumId"]);
-			echo $ux_edit_album_name = html_entity_decode($_REQUEST["edit_album_name"]);
-			$ux_edit_desciption = html_entity_decode(esc_attr($_REQUEST["ux_edit_description"]));
-			$ux_edit_image_width = 160;
-			$ux_edit_image_height = 120;
-			$ux_edit_border_width = 5;
-			$ux_edit_border_color = "#000";
-			$ux_edit_image_no = 3;
-			$ux_edit_slide_interval = 2;
-			$ux_edit_slide = 0;
-			$ux_edit_image_border = 1;
-			
-			$wpdb->query
-			(
-				$wpdb->prepare
-				(
-					"UPDATE " .gallery_bank_albums(). " SET album_name = %s, description = %s, image_width = %d, image_height = %d,
-					border_enable = %d, border_width = %d, border_color = %s, slideshow = %d, slideshow_interval = %d WHERE album_id = %d",
-					$ux_edit_album_name,
-					$ux_edit_desciption,
-					$ux_edit_image_width,
-					$ux_edit_image_height,
-					$ux_edit_image_border,
-					$ux_edit_border_width,
-					$ux_edit_border_color,
-					$ux_edit_slide,
-					$ux_edit_slide_interval,
-					$albumId
-				)
-			);
-			die();
-		}
-		else if($_REQUEST['param'] == "Delete_album")
-		{
-			$album_id = intval($_REQUEST['album_id']);
-			$pics = $wpdb->get_var
-			(
-				$wpdb->prepare
-				(
-					'SELECT count(pic_id) FROM ' . gallery_bank_pics() . ' where album_id =%d', 
-					$album_id
-				)
-			);
-			if($pics != 0)
+			$setting = "image_size:1;width:160px;height:120px;images_in_row:3;image_opacity:1;image_border_size:2px;image_border_radius:2px;image_border_color:rgb(0, 0, 0);/cover_size:1;width:160px;height:120px;cover_opacity:1;cover_border_size:2px;cover_border_radius:2px;border_color:rgb(0, 0, 0);/overlay_opacity:0.6;overlay_border_size:0px;overlay_border_radius:0px;text_color:rgb(0, 0, 0);overlay_border_color:rgb(255, 255, 255);inline_bg_color:rgb(255, 255, 255);overlay_bg_color:rgb(0, 0, 0);/autoplay:0;slide_interval:2;/pagination:0;";
+			if($cover == "undefined")
 			{
-				echo "Pictures Exist";
+				$wpdb->query
+				(
+					$wpdb->prepare
+					(
+						"INSERT INTO ".gallery_bank_settings()."(album_id, album_settings, setting_content)
+						VALUES(%d, %d, %s)",
+						$EventLastId,
+						1,
+						$setting
+					)
+				);
 			}
 			else
 			{
@@ -102,11 +48,82 @@ else
 				(
 					$wpdb->prepare
 					(
-						"DELETE FROM ".gallery_bank_albums()." WHERE album_id = %d",
-						$album_id
+						"INSERT INTO ".gallery_bank_settings()."(album_id, album_settings, setting_content, album_cover)
+						VALUES(%d, %d, %s, %s)",
+						$EventLastId,
+						1,
+						$setting,
+						$cover
 					)
 				);
 			}
+			die();
+		}
+		elseif($_REQUEST["param"] == "update_album")
+		{
+			$albumId = intval($_REQUEST["albumId"]);
+			$ux_edit_album_name = html_entity_decode(esc_attr($_REQUEST["edit_album_name"]));
+			$ux_edit_desciption = html_entity_decode(esc_attr($_REQUEST["ux_edit_description"]));
+			$edit_cover = esc_attr($_REQUEST["cover_array"]);
+			$wpdb->query
+			(
+				$wpdb->prepare
+				(
+					"UPDATE " .gallery_bank_albums(). " SET album_name = %s, description = %s WHERE album_id = %d",
+					$ux_edit_album_name,
+					$ux_edit_desciption,
+					$albumId
+				)
+			);
+			$setting = "image_size:1;width:160px;height:120px;images_in_row:3;image_opacity:1;image_border_size:2px;image_border_radius:2px;image_border_color:rgb(0, 0, 0);/cover_size:1;width:160px;height:120px;cover_opacity:1;cover_border_size:2px;cover_border_radius:2px;border_color:rgb(0, 0, 0);/overlay_opacity:0.6;overlay_border_size:0px;overlay_border_radius:0px;text_color:rgb(0, 0, 0);overlay_border_color:rgb(255, 255, 255);inline_bg_color:rgb(255, 255, 255);overlay_bg_color:rgb(0, 0, 0);/autoplay:0;slide_interval:2;/pagination:0;";
+			if($edit_cover == "undefined")
+			{
+				$wpdb->query
+				(
+					$wpdb->prepare
+					(
+						"UPDATE ".gallery_bank_settings()." SET album_settings = %d ,setting_content = %s WHERE album_id = %d",
+						1,
+						$setting,
+						$albumId
+					)
+				);
+			}
+			else
+			{
+				$wpdb->query
+				(
+					$wpdb->prepare
+					(
+						"UPDATE ".gallery_bank_settings()." SET album_cover = %s, album_settings = %d ,setting_content = %s WHERE album_id = %d",
+						$edit_cover,
+						1,
+						$setting,
+						$albumId
+					)
+				);
+			}
+			die();
+		}
+		else if($_REQUEST['param'] == "Delete_album")
+		{
+			$album_id = intval($_REQUEST['album_id']);
+			$wpdb->query
+			(
+				$wpdb->prepare
+				(
+					"DELETE FROM ".gallery_bank_albums()." WHERE album_id = %d",
+					$album_id
+				)
+			);
+			$wpdb->query
+			(
+				$wpdb->prepare
+				(
+					"DELETE FROM ".gallery_bank_settings()." WHERE album_id = %d",
+					$album_id
+				)
+			);
 			die();
 		}
 		else if($_REQUEST["param"] == "add_pic")
@@ -116,20 +133,27 @@ else
 			$ux_title = html_entity_decode($_REQUEST["title"]);
 			$ux_detail = html_entity_decode($_REQUEST["detail"]);
 			$thumbnail_url = esc_attr($_REQUEST["thumb"]);
-			$alb_cover = intval($_REQUEST["alb_cover"]);
 			$wpdb->query
 			(
 				$wpdb->prepare
 				(
-					"INSERT INTO ".gallery_bank_pics()."(album_id,pic_path,thumbnail_url,title,description,date,album_cover)
-					VALUES(%d,%s,%s,%s,%s,CURDATE(),%d)",
+					"INSERT INTO ".gallery_bank_pics()."(album_id,pic_path,thumbnail_url,title,description,date)
+					VALUES(%d,%s,%s,%s,%s,CURDATE())",
 					$ux_albumid,
 					$ux_path,
 					$thumbnail_url,
 					$ux_title,
-					$ux_detail,
-					$alb_cover
-					
+					$ux_detail
+				)
+			);
+			$pic_id = $wpdb->insert_id;
+			$wpdb->query
+			(
+				$wpdb->prepare
+				(
+					"UPDATE " .gallery_bank_pics(). " SET sorting_order = %d WHERE pic_id = %d",
+					$pic_id,
+					$pic_id
 				)
 			);
 			die();
@@ -140,15 +164,13 @@ else
 			$albumId = intval($_REQUEST['albumId']);
 			$edit_title = html_entity_decode($_REQUEST['edit_title']);
 			$edit_detail = html_entity_decode($_REQUEST['edit_detail']);
-			$edit_cover = esc_attr($_REQUEST["edit_cover"]);
 			$wpdb->query
 			(
 				$wpdb->prepare
 				(
-					"UPDATE " .gallery_bank_pics(). " SET title = %s, description= %s, album_cover = %d WHERE pic_id = %d",
+					"UPDATE " .gallery_bank_pics(). " SET title = %s, description= %s WHERE pic_id = %d",
 					$edit_title,
 					$edit_detail,
-					$edit_cover,
 					$pic_id
 				)
 			);
@@ -159,7 +181,7 @@ else
 		{
 			$picture_id = $_REQUEST['id'];
 			$count_pic = intval($_REQUEST['count_pic']);
-			$Pic_id =  (explode(",",$picture_id));
+			$pic_id =  (explode(",",$picture_id));
 			for($flag = 0; $flag < $count_pic; $flag++ )
 			{
 				$wpdb->query
@@ -167,12 +189,42 @@ else
 					$wpdb->prepare
 					(
 						"DELETE FROM ".gallery_bank_pics()." WHERE pic_id = %d",
-						$Pic_id[$flag]
+						$pic_id[$flag]
 					)
 				);
 			}
 			die();
 		}
+		else if($_REQUEST['param'] == "delete_album")
+		{
+			$album_id = intval($_REQUEST['album_id']);
+			$wpdb->query
+			(
+				$wpdb->prepare
+				(
+					"DELETE FROM ".gallery_bank_pics()." WHERE album_id = %d",
+					$album_id
+				)
+			);
+			$wpdb->query
+			(
+				$wpdb->prepare
+				(
+					"DELETE FROM ".gallery_bank_settings()." WHERE album_id = %d",
+					$album_id
+				)
+			);
+			$wpdb->query
+			(
+				$wpdb->prepare
+				(
+					"DELETE FROM ".gallery_bank_albums()." WHERE album_id = %d",
+					$album_id
+				)
+			);
+			die();
+		}
+		
 		else if($_REQUEST["param"] == "reorderControls")
 		{
 			
@@ -191,6 +243,13 @@ else
 				);
 				$listingCounter = $listingCounter + 1;	
 			}
+			die();
+		}
+		else if($_REQUEST['param'] == "delete_all_albums")
+		{
+			include_once 'gallery-bank.php';
+			plugin_uninstall_script_for_gallery_bank();
+			plugin_install_script_for_gallery_bank();
 			die();
 		}
 	}
